Закончил 4, 5, 6, 7 разделы

+++ project/Java programming learning platform/src/main/java/by/bsac/controller/ConstructorController.java
@@ -109,8 +109,10 @@ public class ConstructorController {
                 }
             }
         }
-        File imgCourse = new File(this.pathImgCourse, course.getImgName());
-        imgCourse.delete();
+        if (!StringUtils.isEmpty(course.getImgName())) {
+            File imgCourse = new File(this.pathImgCourse, course.getImgName());
+            imgCourse.delete();
+        }
         courseRepository.delete(course);
         return "redirect:/";
     }
